Rough-in handler loading
Going to try using dlsym directly on the tcmu_handler struct that each
handler must have, instead of calling a registration function. Simpler?
Unless there's a reason everyone does it the other way, in which case
we'll find out shortly :)

Currently searching in the current directory for files named "handler_*".

Signed-off-by: Andy Grover <agrover@redhat.com>